Add homepage links to sample data: #67, #73
Add all links that appear on the live DCSOps homepage to the database.
Note that I expect that these should vary with the logged-in user's
chapter.  I didn't include the real URLs.  The only one that is within
DCSOps is the dispatch console, which *definitely* varies by chapter and
which I'll look into next.

Note that I expect the visibility of these links will vary with user
privilege level, but I haven't learned in what way just yet.