Link correctly to dispatch console: #67, #73
Add an incident scope to the sample data so that the dispatch console
will work correctly.  Also, fix the console link from the homepage so
that it can be tested.

Links from the homepage that are still broken and should work:
scheduler, incidents.  I'll look at those next.

I do not expect links to external services or files to work locally at
this stage, if ever.