Add day and night times; create sample shifts: #67
Add a migration to create the start of day and night for every chapter.
I think that the usage in `flex_schedules_helper.rb` suggests that these
columns should exist, and adding them makes the tooltips show up in
`app/views/scheduler/home/root.html.haml` as expected.

Add shifts to sample data so that the schedule loads correctly.

Note that I applied the migration to my own database and so db/schema.rb
changed.  Since we have been keeping this file under version control I'm
committing this change, but I suspect we should ignore it, since the
changes are auto-generated.