All homepage links now work: #73
Undo changes from 2f25ea8, now that I understand that
incidents_chapter_root_path expected to pull the url_slug component out
of a Chapter object.  Add a value for url_slug to our test Chapter
object, accordingly.  Add an incident scope for the chapter, so that the
Incidents link in the header and first set of homepage links goes
somewhere.

Note that the incidents page should have a "Submit Incident Report" link
and does not, presumably because of a permissions conditional in
`app/views/incidents/home/root.html.haml`.  I'll look into that next and
see how to make the test user an admin user.