Load map of recent incidents correctly: #67, #73
Make incidents map load correctly.  Create two incidents, with validate:
false for now because I'm not sure how to correctly validate incident
numbers.  Note that I need to set incident dates so that they are always
in the past two weeks from the time that the sample data is loaded (so
that they will show up in the "Recent Incidents" list and on the map).
Use chapter config values from a FactoryGirl file, so that presumably
all information is now included and the map boundaries exist.

Change the way the map boundaries are loaded in map_helper.rb.  I'm not
sure why the old way wasn't fetching them correctly, but the new way
brings them in from the database as expected.  It might be less
Rubyesque.  The fact that the other way didn't work probably also points
to some missing data link that I will find as I explore.