Create admin user: #73, #67
Grant more privileges to a user, and rename to example_admin.  Now we
have two users being created in the sample data.  One of them is an
admin who has the authority to submit incident reports and create new
incidents.  The other is a "regular" user without any special
privileges.  I expect the "regular" user to act like a volunteer.  That
user should be able to sign up for shifts and be assigned to incidents.
The admin user should be able to create incidents and dispatch
volunteers to them.

Next is to create more regions, so that all US addresses are valid for
new incidents.  Also, check through the close incident workflow to make
sure that the admin user has all the necessary powers.

Our admin user may need to have access to more chapters.  So far the
admin only has access to the Bay Area chapter, since that's the only one
created in our sample data.