Show "responders" console: #73, #67
Each chapter needs to have "incidents_enable_dispatch_console" set to
true in order for users to be able to add responders.  This came up
while I was researching issue #70.

The responders console now shows up, but parts of it are still blocked
based on the "dispatch_config" setting, which is related to the
"territory" of the incident.  Next commit will address that.