Sync up code differences between Heroku and GitHub
The git commit SHA currently running on Heroku was not present in the
source tree of the repo on GitHub (probably the result of a local
change pushed to Heroku but not to GitHub).

This commit brings the changes extant on Heroku into the main repo. As
you can see the changes are minimal so the purpose of this is mostly to
document that it was done, rather than to preserve the inconsequential
changes.

These changes were retrieved by downloading the currently-running
source code slug from Heroku (via
https://github.com/heroku/heroku-slugs), copying the slug code over the
repo code, and committing the difference.