Fix database seeds, round 1
The database seeds appear to be out of date with the current models and
schema, since they do not load cleanly due to missing columns,
associations, etc.

This brings the position role seeds up to date by converting the
position `grants_role` attribute into the models that attribute
eventually became.

The procedure I followed to do this and make sure the outcome was as
intended was:

1. Start with an empty database
2. Run migrations up to the point at which these seeds were created
3. Run seeds (while hacking a couple fixes to let seeds run
   successfully (unrelated to roles structure)
4. Run remaining migrations to migrate original role attributes into
   final form
5. Inspect and save resulting records for comparison
6. Add seed code that will create same records from scratch under
   current models and schema.

After that, running seeds on a fresh database works. Well, up to a
point; more work needs to be done to fix problems further down the
seeds file, which will come in a subsequent commit.