Fix database seeds, round 2
The database seeds appear to be out of date with the current models and
schema, since they do not load cleanly due to missing columns,
associations, etc.

This brings the shift seeds up to date by changing various attributes
and associations to match the current ones. I was able to piece
together what needs to go where by examining past migrations and how
they transformed the data, along with the current models.