Require `url_slug` attribute of Chapters
Some obscure error is happening in the rails routing code when url_slug
is not present and `Roster::Chapter#to_param` falls back to using the
ID. I suspect a bad interaction with inherited_resources.

The error message is `undefined method `empty?' for 1:Fixnum` (1 being
a chapter ID), which comes from calling
`incidents_chapter_root_path(current_chapter)` in a view template. The
rest of the stack trace is pretty useless aside from pointing to the
routing/URL helper code.

In any case, a quick and expedient resolution without digging further
is just to make sure chapters have URL slugs. So I added a model
validation for that field and added one to the seed chapter. Production
chapters all have an URL slug already so this is a safe validation to
add.