Spec fixes: remove randomness from url_slug
Randomness in tests isn't the best. In this case it's a problem because
rspec seeds the RNG with a known value for repeatable randomness across
test runs, which conflicts when creating multiple chapters from
factories due to the url_slug not being unique.

I don't see any reason to generate a random slug for these; we should
probably just use normal Factory Girl sequences if we need uniqueness.