Setup separate staging environment for deployments
Updates environment settings so that a staging environment can be run on
Heroku with settings similar to production, but without services like
email that will contact users or change production data. Creates a rake
task for updating the staging database with a truncated copy of the most
recent production backup.