Remove call to VC when logging in in legacy mode.
Legacy mode was doing two things, attempting to authenticate against the
database, and then making a call out to Volunteer Connection.  This
second one stopped working when SSO was implemented, and indeed the call
returned a bunch of html.

This removes the call as a first step toward cutting that entire portion
of the app out.