Fix tests on travis, login avoids redirect
For some reason, travis capybara doesn't follow redirects on the 1+n
times it hits a page, so going to "/" was just failing on every test
after the first one, when wanting to login.  So instead, making the
login code go right to the correct place was a simple fix.

Also include logging out after every feature call, as well as some
whitespace fixes while in there.