Remove seeds authlogic hack
This was causing the seeds not to load due to a failed validation on
the password confirmation field, which should not be validated.
Authlogic is configured in the Person model to disable it but this hack
did not faithfully copy that configuration, causing the failure.

As an experiment I tried seeding with this commented out and it
succeeded, so I don't think it's still necessary.