Additional capybara/apparition config
This disables JS exceptions that occur in the browser during tests
being re-raised in Ruby.

Arguably we should care about JS errors occurring, however the existing
behavior under capybara-webkit was to not care, so this maintains that.

It also removes some commented Sauce config which I think is a visual
testing tool that would be replaced by this.