Rails 5 upgrade: upgrade roadie
Roadie has evolved to have its Rails-specific components extracted into
roadie-rails, so we switch to using that and update to the latest
version.

roadie-rails make the asset provider config by default, so we can
remove that line. The other necessary change is that roadie no longer
automatically configures all mailers, so we turn it on explicitly in
MailerCommon which is included by all mailers.