Add spec for CleanEmptyStrings
The intent of this code seems to be to replace empty strings with nils
on save (or specifically, on type case for DB writes).

The way it works no longer exists in Rails 5, so I'll be finding a
workaround for that case. In the meantime, this makes the behavior
explicit and tested so that we'll be able to catch regressions.