Rails 5 upgrade: run Rails app update script
Rails ships with a script that will write new default files that would
have been created if the app was started on the current version [1].

This commit runs that script and merges our custom config/files with
the new Rails 5.2 default config/files.

It also removes the rails_12factor gem, which no longer needed on
Rails 5 [2] with the updated config options added in this commit.

This brings the app more inline with current Rails.

[1] https://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#the-update-task

[2] https://github.com/heroku/rails_12factor#migrating-to-rails-5