Remove method_missing in ApplicationHelper
This appears to be left over from when this code was part of an engine?
Because the only place `main_app` exists in Rails is to allow an Engine
to access the app it's included in. So it doesn't make sense that this
app, not being an Engine, calls that.

In any case, having a method_missing defined in ApplicationHelper
obfuscates a TON of errors, any method call in any view.

I'm removing this and if it turns out it was indeed used we can
consider adding it back, or better yet, only adding back the methods we
actually need.