Change County to Shift Territory
This change is a bit more wide ranging than the other renames, because
counties sometimes mean actual physical counties, and other times mean
a shift territory that can be a city, zip, county, or collection of
counties.

For that reason, many of the times, county was changed to
shift_territory, but there are many cases where a county_name became a
county since there was no more confusion.

Issue #185: Change entity names