Rename Roster::Role to Capability
This does not affect other roles such as homepage link roles, and
notification roles, but only the roles attached to positions.  These
weren't actually roles, so the name brings them more in line with their
actual purpose (capabilities), and reduces confusion.

This is another wide, but not deep, change.

Issue #185: Change entity names