SerializedColumns Migration: Step 2.2
Continuing the process of replacing the use of Postgres `hstore`
columns with `jsonb` columns, which Rails has native type casting
support for. Then we'll be able to remove the custom type casting code
from SerializedColumns which has tight coupling into AR internals.

This commit is the second half of Step 2: update SerializedColumns
module to drop hstore-specific type casting machinery and adjust any
call sites as necessary.

Mainly this changes the source for the arcdata_core gem (where
SerializedColumns lives) to the branch that has all the changes for
using JSONB columns for storage. The bulk of the code changes for this
step can be found there.

The API is as close to unchanged as possible, so the necessary changes in this app were minimal:

- A couple factories where previously we set the storage field with a
  full hash of serialized values, rather than calling the setter method
  provided by SerializedColumns. This caused some problems by skipping
  the type casting in the setter and using symbol hash keys when
  elsewhere we expect strings (as they would be if returned from the DB)

- One place where we ran a raw query against the serialized column
  using a special hstore operator (changed to the equivalent JSONB
  operator).