Fix #224: Add an index to versions
The Current Activity link was slow and in production the versions
lookup was taking 5-40 seconds.  In development, it was taking 1.5
seconds, and this index improved that to 1.5 milliseconds.

Some other db/schema.rb changes came along for the ride, because of
rails adding stuff when running db:migrate.  This should slow down
after the next few migrations.

Issue #224: Fix the "Current Activity" link.