Fix #201: Allow incidents to be closed without completion
For admins, add a buton that forces an incident to close without fully
submitting all the validating information needed for it.

This also adds a small change, adding a link from Currently Open
Incidents to the incident pages for easier navigating.

Issue #201: Create a way for closing open incident reports listed
            on the Currently Open Incidents page without having to
            complete all mandatory fields