Fix #205: Add reason_marked_invalid to incident
Before, the reason an incident was marked invalid was stored in the
incident_type field, which resulted in losing the original incident type
(fire, or vacate, or whatever).  This caused issues when the incident
was reopened, or for reporting.

Instead, we want to have a separate field for the invalid reason that we
display and search on when the incident status is 'invalid'

Issue #205: Create a Response Status field separate from Incident Types