Create realignment script for illinois
This creates the script and necessary DB changes to do the DCSOps
realignment.  It creates the new concept "VC Position", and then
populates the database with data from the Illinois spreadsheet.

This is a first pass, and will be improved and added to.

Issue #255: Write scripts from Red Cross-provided spreadsheets and
            import regional configuration data into DCSOps