Update vc importer code for new vc queries
The Volunteer Connection Query Tool moved from having an import per
region, with a different login for each region to having grouped imports
accessible to one account.  This was also a reduced spreadsheet as a lot
of information on old spreadsheets was not required.

This pulls the data, groups the users by the regions, and then processes
as it did before.  This is part of the red cross realignment project.

Issue #255: Write scripts from Red Cross-provided spreadsheets
            and import regional configuration data into DCSOps