Fix #36/#253 part 2: Instead of polygon, use kmz file
Because the polygons representing boundaries were too large to input
manually, we need to move to a system where we can use a configured kmz
file somewhere on the internet with the kml layer.  For demonstration
purposes, we are checking in Illinois, but we may remove later and host
them elsewhere.

Issue #253: Move scope map configuration from fusion tables to inline polygon
Issue #36: Update Regional boundary maps on Region's Incidents pages