Fix #211 part 2: remove incidents from dispatch console
This removes incidents from the dispatch console if they have been
dispatched through the normal incident interface.  That happens when a
timeentry is added in the "Incident Dispatched" type.  Also a small
refactor where a DRY violation was removed with the new incident method.

Issue #211: On the Disaster Dispatch page accessed via clicking the
            NorCen Dispatch Console homepage button and then the Dispatch
            Console button, create a way to delete the incidents listed
            under the Dispatching Incidents heading