Base notification roles on position and shift territory
Before, they were only based on positions, which was problematic under
realignment because positions are more general now.  Before, positions
in GNY, for instance, had things like "DAT Supervisor - Bronx" and so we
could create a notification role for just that.  Now, we have a "DAT
Supervisor" position for the "Bronx" shift territory.

Notification roles needed to be updated to reflect this.

Issue #255: Write scripts from Red Cross-provided spreadsheets and
            import regional configuration data into DCSOps