Add id to notification roles configuration
Because we changed the table from a previous one, where the id was not
present in the table, active admin was throwing errors when trying to
update things.  This is due to how we used a different part of
activeadmin to generate the position/shift territory sync up to the
notifications roles.